REFACTOR: Split Monolithic Zustand Store
**Current:** index.ts — 1070 lines, 50+ actions, types, and mappers in one file.
**Future:** Domain-oriented slices; same Zustand instance, same `useAppStore()` API.

11. Extract store types
12. Extract mapper functions
13. Extract chat actions (message mutations, tree ops, branching)
14. Extract entity CRUD actions
15. Slim index.ts to ~50 lines (compose slices, re-export hook + types)
