+++ app/(main)/chats/page.tsx
@@ -1,6 +1,6 @@
 import { listChats } from "@/lib/actions/chats/list-chats";
 import { ChatsClient } from "./_components/chats-client";
-import { chatRowToStore } from "@/lib/store";
+import { chatRowToStore } from "@/lib/store/mappers/chat";
 import type { ChatRow } from "@/types/chat-row";
 
 export default async function ChatsPage() {
