+++ components/chat/chat-card.tsx
@@ -4,7 +4,7 @@ import { Card } from "@/components/ui/card";
 import { MessageSquare } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { ROUTES } from "@/lib/routes";
-import type { Chat } from "@/lib/store";
+import type { Chat } from "@/types/chat";
 import { ChatOptions } from "./chat-options";
 
 /**
