+++ components/knowledgebase/knowledgebase-options.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Trash2, Edit2 } from "lucide-react";
 import { useIsMobile } from "@/hooks/use-mobile";
 import { useAppStore } from "@/lib/store";
-import type { Knowledgebase } from "@/lib/store";
+import type { Knowledgebase } from "@/types/knowledgebase";
 import { RenameDialog } from "@/components/shared/rename-dialog";
 import { ResponsiveMenu } from "@/components/shared/responsive-menu";
 import { toast } from "sonner";
