+++ types/attachment.ts
@@ -1,0 +1,14 @@
+/**
+ * A file attachment on a message — images and documents.
+ */
+export type Attachment = {
+  id: string;
+  type: "image" | "document";
+  name: string;
+  mimeType: string;
+  sizeBytes: number;
+  dataUrl: string;
+  extractedText?: string;
+  key?: string;
+  url?: string;
+};
